@@ -1,6 +1,6 @@
 // Code generated by protoc-gen-go. DO NOT EDIT.
 // versions:
-// 	protoc-gen-go v1.28.1
+// 	protoc-gen-go v1.35.2
 // 	protoc        v3.20.0
 // source: common/net.proto
 
@@ -34,11 +34,9 @@ type IPAddress struct {
 
 func (x *IPAddress) Reset() {
 	*x = IPAddress{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_common_net_proto_msgTypes[0]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_common_net_proto_msgTypes[0]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *IPAddress) String() string {
@@ -49,7 +47,7 @@ func (*IPAddress) ProtoMessage() {}
 
 func (x *IPAddress) ProtoReflect() protoreflect.Message {
 	mi := &file_common_net_proto_msgTypes[0]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -112,11 +110,9 @@ type IPv6 struct {
 
 func (x *IPv6) Reset() {
 	*x = IPv6{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_common_net_proto_msgTypes[1]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_common_net_proto_msgTypes[1]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *IPv6) String() string {
@@ -127,7 +123,7 @@ func (*IPv6) ProtoMessage() {}
 
 func (x *IPv6) ProtoReflect() protoreflect.Message {
 	mi := &file_common_net_proto_msgTypes[1]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -167,11 +163,9 @@ type TcpAddress struct {
 
 func (x *TcpAddress) Reset() {
 	*x = TcpAddress{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_common_net_proto_msgTypes[2]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_common_net_proto_msgTypes[2]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *TcpAddress) String() string {
@@ -182,7 +176,7 @@ func (*TcpAddress) ProtoMessage() {}
 
 func (x *TcpAddress) ProtoReflect() protoreflect.Message {
 	mi := &file_common_net_proto_msgTypes[2]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -250,7 +244,7 @@ func file_common_net_proto_rawDescGZIP() []byte {
 }
 
 var file_common_net_proto_msgTypes = make([]protoimpl.MessageInfo, 3)
-var file_common_net_proto_goTypes = []interface{}{
+var file_common_net_proto_goTypes = []any{
 	(*IPAddress)(nil),  // 0: linkerd2.common.net.IPAddress
 	(*IPv6)(nil),       // 1: linkerd2.common.net.IPv6
 	(*TcpAddress)(nil), // 2: linkerd2.common.net.TcpAddress
@@ -270,45 +264,7 @@ func file_common_net_proto_init() {
 	if File_common_net_proto != nil {
 		return
 	}
-	if !protoimpl.UnsafeEnabled {
-		file_common_net_proto_msgTypes[0].Exporter = func(v interface{}, i int) interface{} {
-			switch v := v.(*IPAddress); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-		file_common_net_proto_msgTypes[1].Exporter = func(v interface{}, i int) interface{} {
-			switch v := v.(*IPv6); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-		file_common_net_proto_msgTypes[2].Exporter = func(v interface{}, i int) interface{} {
-			switch v := v.(*TcpAddress); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-	}
-	file_common_net_proto_msgTypes[0].OneofWrappers = []interface{}{
+	file_common_net_proto_msgTypes[0].OneofWrappers = []any{
 		(*IPAddress_Ipv4)(nil),
 		(*IPAddress_Ipv6)(nil),
 	}
