@@ -1,6 +1,6 @@
 // Code generated by protoc-gen-go. DO NOT EDIT.
 // versions:
-// 	protoc-gen-go v1.28.1
+// 	protoc-gen-go v1.35.2
 // 	protoc        v3.20.0
 // source: viz.proto
 
@@ -191,11 +191,9 @@ type Empty struct {
 
 func (x *Empty) Reset() {
 	*x = Empty{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_viz_proto_msgTypes[0]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_viz_proto_msgTypes[0]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *Empty) String() string {
@@ -206,7 +204,7 @@ func (*Empty) ProtoMessage() {}
 
 func (x *Empty) ProtoReflect() protoreflect.Message {
 	mi := &file_viz_proto_msgTypes[0]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -234,11 +232,9 @@ type CheckResult struct {
 
 func (x *CheckResult) Reset() {
 	*x = CheckResult{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_viz_proto_msgTypes[1]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_viz_proto_msgTypes[1]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *CheckResult) String() string {
@@ -249,7 +245,7 @@ func (*CheckResult) ProtoMessage() {}
 
 func (x *CheckResult) ProtoReflect() protoreflect.Message {
 	mi := &file_viz_proto_msgTypes[1]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -300,11 +296,9 @@ type SelfCheckRequest struct {
 
 func (x *SelfCheckRequest) Reset() {
 	*x = SelfCheckRequest{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_viz_proto_msgTypes[2]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_viz_proto_msgTypes[2]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *SelfCheckRequest) String() string {
@@ -315,7 +309,7 @@ func (*SelfCheckRequest) ProtoMessage() {}
 
 func (x *SelfCheckRequest) ProtoReflect() protoreflect.Message {
 	mi := &file_viz_proto_msgTypes[2]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -340,11 +334,9 @@ type SelfCheckResponse struct {
 
 func (x *SelfCheckResponse) Reset() {
 	*x = SelfCheckResponse{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_viz_proto_msgTypes[3]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_viz_proto_msgTypes[3]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *SelfCheckResponse) String() string {
@@ -355,7 +347,7 @@ func (*SelfCheckResponse) ProtoMessage() {}
 
 func (x *SelfCheckResponse) ProtoReflect() protoreflect.Message {
 	mi := &file_viz_proto_msgTypes[3]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -387,11 +379,9 @@ type ListServicesRequest struct {
 
 func (x *ListServicesRequest) Reset() {
 	*x = ListServicesRequest{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_viz_proto_msgTypes[4]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_viz_proto_msgTypes[4]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *ListServicesRequest) String() string {
@@ -402,7 +392,7 @@ func (*ListServicesRequest) ProtoMessage() {}
 
 func (x *ListServicesRequest) ProtoReflect() protoreflect.Message {
 	mi := &file_viz_proto_msgTypes[4]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -434,11 +424,9 @@ type ListServicesResponse struct {
 
 func (x *ListServicesResponse) Reset() {
 	*x = ListServicesResponse{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_viz_proto_msgTypes[5]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_viz_proto_msgTypes[5]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *ListServicesResponse) String() string {
@@ -449,7 +437,7 @@ func (*ListServicesResponse) ProtoMessage() {}
 
 func (x *ListServicesResponse) ProtoReflect() protoreflect.Message {
 	mi := &file_viz_proto_msgTypes[5]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -482,11 +470,9 @@ type Service struct {
 
 func (x *Service) Reset() {
 	*x = Service{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_viz_proto_msgTypes[6]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_viz_proto_msgTypes[6]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *Service) String() string {
@@ -497,7 +483,7 @@ func (*Service) ProtoMessage() {}
 
 func (x *Service) ProtoReflect() protoreflect.Message {
 	mi := &file_viz_proto_msgTypes[6]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -536,11 +522,9 @@ type ListPodsRequest struct {
 
 func (x *ListPodsRequest) Reset() {
 	*x = ListPodsRequest{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_viz_proto_msgTypes[7]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_viz_proto_msgTypes[7]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *ListPodsRequest) String() string {
@@ -551,7 +535,7 @@ func (*ListPodsRequest) ProtoMessage() {}
 
 func (x *ListPodsRequest) ProtoReflect() protoreflect.Message {
 	mi := &file_viz_proto_msgTypes[7]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -583,11 +567,9 @@ type ListPodsResponse struct {
 
 func (x *ListPodsResponse) Reset() {
 	*x = ListPodsResponse{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_viz_proto_msgTypes[8]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_viz_proto_msgTypes[8]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *ListPodsResponse) String() string {
@@ -598,7 +580,7 @@ func (*ListPodsResponse) ProtoMessage() {}
 
 func (x *ListPodsResponse) ProtoReflect() protoreflect.Message {
 	mi := &file_viz_proto_msgTypes[8]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -649,11 +631,9 @@ type Pod struct {
 
 func (x *Pod) Reset() {
 	*x = Pod{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_viz_proto_msgTypes[9]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_viz_proto_msgTypes[9]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *Pod) String() string {
@@ -664,7 +644,7 @@ func (*Pod) ProtoMessage() {}
 
 func (x *Pod) ProtoReflect() protoreflect.Message {
 	mi := &file_viz_proto_msgTypes[9]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -859,11 +839,9 @@ type HttpMethod struct {
 
 func (x *HttpMethod) Reset() {
 	*x = HttpMethod{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_viz_proto_msgTypes[10]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_viz_proto_msgTypes[10]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *HttpMethod) String() string {
@@ -874,7 +852,7 @@ func (*HttpMethod) ProtoMessage() {}
 
 func (x *HttpMethod) ProtoReflect() protoreflect.Message {
 	mi := &file_viz_proto_msgTypes[10]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -940,11 +918,9 @@ type Scheme struct {
 
 func (x *Scheme) Reset() {
 	*x = Scheme{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_viz_proto_msgTypes[11]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_viz_proto_msgTypes[11]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *Scheme) String() string {
@@ -955,7 +931,7 @@ func (*Scheme) ProtoMessage() {}
 
 func (x *Scheme) ProtoReflect() protoreflect.Message {
 	mi := &file_viz_proto_msgTypes[11]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -1017,11 +993,9 @@ type Headers struct {
 
 func (x *Headers) Reset() {
 	*x = Headers{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_viz_proto_msgTypes[12]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_viz_proto_msgTypes[12]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *Headers) String() string {
@@ -1032,7 +1006,7 @@ func (*Headers) ProtoMessage() {}
 
 func (x *Headers) ProtoReflect() protoreflect.Message {
 	mi := &file_viz_proto_msgTypes[12]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -1068,11 +1042,9 @@ type Eos struct {
 
 func (x *Eos) Reset() {
 	*x = Eos{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_viz_proto_msgTypes[13]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_viz_proto_msgTypes[13]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *Eos) String() string {
@@ -1083,7 +1055,7 @@ func (*Eos) ProtoMessage() {}
 
 func (x *Eos) ProtoReflect() protoreflect.Message {
 	mi := &file_viz_proto_msgTypes[13]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -1145,11 +1117,9 @@ type ApiError struct {
 
 func (x *ApiError) Reset() {
 	*x = ApiError{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_viz_proto_msgTypes[14]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_viz_proto_msgTypes[14]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *ApiError) String() string {
@@ -1160,7 +1130,7 @@ func (*ApiError) ProtoMessage() {}
 
 func (x *ApiError) ProtoReflect() protoreflect.Message {
 	mi := &file_viz_proto_msgTypes[14]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -1192,11 +1162,9 @@ type PodErrors struct {
 
 func (x *PodErrors) Reset() {
 	*x = PodErrors{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_viz_proto_msgTypes[15]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_viz_proto_msgTypes[15]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *PodErrors) String() string {
@@ -1207,7 +1175,7 @@ func (*PodErrors) ProtoMessage() {}
 
 func (x *PodErrors) ProtoReflect() protoreflect.Message {
 	mi := &file_viz_proto_msgTypes[15]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -1251,11 +1219,9 @@ type Resource struct {
 
 func (x *Resource) Reset() {
 	*x = Resource{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_viz_proto_msgTypes[16]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_viz_proto_msgTypes[16]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *Resource) String() string {
@@ -1266,7 +1232,7 @@ func (*Resource) ProtoMessage() {}
 
 func (x *Resource) ProtoReflect() protoreflect.Message {
 	mi := &file_viz_proto_msgTypes[16]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -1319,11 +1285,9 @@ type ResourceSelection struct {
 
 func (x *ResourceSelection) Reset() {
 	*x = ResourceSelection{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_viz_proto_msgTypes[17]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_viz_proto_msgTypes[17]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *ResourceSelection) String() string {
@@ -1334,7 +1298,7 @@ func (*ResourceSelection) ProtoMessage() {}
 
 func (x *ResourceSelection) ProtoReflect() protoreflect.Message {
 	mi := &file_viz_proto_msgTypes[17]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -1374,11 +1338,9 @@ type ResourceError struct {
 
 func (x *ResourceError) Reset() {
 	*x = ResourceError{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_viz_proto_msgTypes[18]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_viz_proto_msgTypes[18]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *ResourceError) String() string {
@@ -1389,7 +1351,7 @@ func (*ResourceError) ProtoMessage() {}
 
 func (x *ResourceError) ProtoReflect() protoreflect.Message {
 	mi := &file_viz_proto_msgTypes[18]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -1437,11 +1399,9 @@ type StatSummaryRequest struct {
 
 func (x *StatSummaryRequest) Reset() {
 	*x = StatSummaryRequest{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_viz_proto_msgTypes[19]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_viz_proto_msgTypes[19]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *StatSummaryRequest) String() string {
@@ -1452,7 +1412,7 @@ func (*StatSummaryRequest) ProtoMessage() {}
 
 func (x *StatSummaryRequest) ProtoReflect() protoreflect.Message {
 	mi := &file_viz_proto_msgTypes[19]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -1559,11 +1519,9 @@ type StatSummaryResponse struct {
 
 func (x *StatSummaryResponse) Reset() {
 	*x = StatSummaryResponse{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_viz_proto_msgTypes[20]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_viz_proto_msgTypes[20]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *StatSummaryResponse) String() string {
@@ -1574,7 +1532,7 @@ func (*StatSummaryResponse) ProtoMessage() {}
 
 func (x *StatSummaryResponse) ProtoReflect() protoreflect.Message {
 	mi := &file_viz_proto_msgTypes[20]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -1637,11 +1595,9 @@ type AuthzRequest struct {
 
 func (x *AuthzRequest) Reset() {
 	*x = AuthzRequest{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_viz_proto_msgTypes[21]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_viz_proto_msgTypes[21]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *AuthzRequest) String() string {
@@ -1652,7 +1608,7 @@ func (*AuthzRequest) ProtoMessage() {}
 
 func (x *AuthzRequest) ProtoReflect() protoreflect.Message {
 	mi := &file_viz_proto_msgTypes[21]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -1695,11 +1651,9 @@ type AuthzResponse struct {
 
 func (x *AuthzResponse) Reset() {
 	*x = AuthzResponse{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_viz_proto_msgTypes[22]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_viz_proto_msgTypes[22]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *AuthzResponse) String() string {
@@ -1710,7 +1664,7 @@ func (*AuthzResponse) ProtoMessage() {}
 
 func (x *AuthzResponse) ProtoReflect() protoreflect.Message {
 	mi := &file_viz_proto_msgTypes[22]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -1778,11 +1732,9 @@ type BasicStats struct {
 
 func (x *BasicStats) Reset() {
 	*x = BasicStats{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_viz_proto_msgTypes[23]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_viz_proto_msgTypes[23]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *BasicStats) String() string {
@@ -1793,7 +1745,7 @@ func (*BasicStats) ProtoMessage() {}
 
 func (x *BasicStats) ProtoReflect() protoreflect.Message {
 	mi := &file_viz_proto_msgTypes[23]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -1872,11 +1824,9 @@ type TcpStats struct {
 
 func (x *TcpStats) Reset() {
 	*x = TcpStats{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_viz_proto_msgTypes[24]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_viz_proto_msgTypes[24]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *TcpStats) String() string {
@@ -1887,7 +1837,7 @@ func (*TcpStats) ProtoMessage() {}
 
 func (x *TcpStats) ProtoReflect() protoreflect.Message {
 	mi := &file_viz_proto_msgTypes[24]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -1935,11 +1885,9 @@ type TrafficSplitStats struct {
 
 func (x *TrafficSplitStats) Reset() {
 	*x = TrafficSplitStats{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_viz_proto_msgTypes[25]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_viz_proto_msgTypes[25]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *TrafficSplitStats) String() string {
@@ -1950,7 +1898,7 @@ func (*TrafficSplitStats) ProtoMessage() {}
 
 func (x *TrafficSplitStats) ProtoReflect() protoreflect.Message {
 	mi := &file_viz_proto_msgTypes[25]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -2000,11 +1948,9 @@ type ServerStats struct {
 
 func (x *ServerStats) Reset() {
 	*x = ServerStats{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_viz_proto_msgTypes[26]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_viz_proto_msgTypes[26]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *ServerStats) String() string {
@@ -2015,7 +1961,7 @@ func (*ServerStats) ProtoMessage() {}
 
 func (x *ServerStats) ProtoReflect() protoreflect.Message {
 	mi := &file_viz_proto_msgTypes[26]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -2078,11 +2024,9 @@ type StatTable struct {
 
 func (x *StatTable) Reset() {
 	*x = StatTable{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_viz_proto_msgTypes[27]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_viz_proto_msgTypes[27]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *StatTable) String() string {
@@ -2093,7 +2037,7 @@ func (*StatTable) ProtoMessage() {}
 
 func (x *StatTable) ProtoReflect() protoreflect.Message {
 	mi := &file_viz_proto_msgTypes[27]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -2142,11 +2086,9 @@ type EdgesRequest struct {
 
 func (x *EdgesRequest) Reset() {
 	*x = EdgesRequest{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_viz_proto_msgTypes[28]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_viz_proto_msgTypes[28]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *EdgesRequest) String() string {
@@ -2157,7 +2099,7 @@ func (*EdgesRequest) ProtoMessage() {}
 
 func (x *EdgesRequest) ProtoReflect() protoreflect.Message {
 	mi := &file_viz_proto_msgTypes[28]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -2193,11 +2135,9 @@ type EdgesResponse struct {
 
 func (x *EdgesResponse) Reset() {
 	*x = EdgesResponse{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_viz_proto_msgTypes[29]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_viz_proto_msgTypes[29]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *EdgesResponse) String() string {
@@ -2208,7 +2148,7 @@ func (*EdgesResponse) ProtoMessage() {}
 
 func (x *EdgesResponse) ProtoReflect() protoreflect.Message {
 	mi := &file_viz_proto_msgTypes[29]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -2274,11 +2214,9 @@ type Edge struct {
 
 func (x *Edge) Reset() {
 	*x = Edge{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_viz_proto_msgTypes[30]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_viz_proto_msgTypes[30]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *Edge) String() string {
@@ -2289,7 +2227,7 @@ func (*Edge) ProtoMessage() {}
 
 func (x *Edge) ProtoReflect() protoreflect.Message {
 	mi := &file_viz_proto_msgTypes[30]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -2355,11 +2293,9 @@ type TopRoutesRequest struct {
 
 func (x *TopRoutesRequest) Reset() {
 	*x = TopRoutesRequest{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_viz_proto_msgTypes[31]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_viz_proto_msgTypes[31]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *TopRoutesRequest) String() string {
@@ -2370,7 +2306,7 @@ func (*TopRoutesRequest) ProtoMessage() {}
 
 func (x *TopRoutesRequest) ProtoReflect() protoreflect.Message {
 	mi := &file_viz_proto_msgTypes[31]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -2450,11 +2386,9 @@ type TopRoutesResponse struct {
 
 func (x *TopRoutesResponse) Reset() {
 	*x = TopRoutesResponse{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_viz_proto_msgTypes[32]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_viz_proto_msgTypes[32]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *TopRoutesResponse) String() string {
@@ -2465,7 +2399,7 @@ func (*TopRoutesResponse) ProtoMessage() {}
 
 func (x *TopRoutesResponse) ProtoReflect() protoreflect.Message {
 	mi := &file_viz_proto_msgTypes[32]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -2528,11 +2462,9 @@ type RouteTable struct {
 
 func (x *RouteTable) Reset() {
 	*x = RouteTable{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_viz_proto_msgTypes[33]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_viz_proto_msgTypes[33]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *RouteTable) String() string {
@@ -2543,7 +2475,7 @@ func (*RouteTable) ProtoMessage() {}
 
 func (x *RouteTable) ProtoReflect() protoreflect.Message {
 	mi := &file_viz_proto_msgTypes[33]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -2582,11 +2514,9 @@ type GatewaysTable struct {
 
 func (x *GatewaysTable) Reset() {
 	*x = GatewaysTable{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_viz_proto_msgTypes[34]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_viz_proto_msgTypes[34]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *GatewaysTable) String() string {
@@ -2597,7 +2527,7 @@ func (*GatewaysTable) ProtoMessage() {}
 
 func (x *GatewaysTable) ProtoReflect() protoreflect.Message {
 	mi := &file_viz_proto_msgTypes[34]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -2631,11 +2561,9 @@ type GatewaysRequest struct {
 
 func (x *GatewaysRequest) Reset() {
 	*x = GatewaysRequest{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_viz_proto_msgTypes[35]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_viz_proto_msgTypes[35]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *GatewaysRequest) String() string {
@@ -2646,7 +2574,7 @@ func (*GatewaysRequest) ProtoMessage() {}
 
 func (x *GatewaysRequest) ProtoReflect() protoreflect.Message {
 	mi := &file_viz_proto_msgTypes[35]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -2696,11 +2624,9 @@ type GatewaysResponse struct {
 
 func (x *GatewaysResponse) Reset() {
 	*x = GatewaysResponse{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_viz_proto_msgTypes[36]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_viz_proto_msgTypes[36]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *GatewaysResponse) String() string {
@@ -2711,7 +2637,7 @@ func (*GatewaysResponse) ProtoMessage() {}
 
 func (x *GatewaysResponse) ProtoReflect() protoreflect.Message {
 	mi := &file_viz_proto_msgTypes[36]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -2783,11 +2709,9 @@ type Headers_Header struct {
 
 func (x *Headers_Header) Reset() {
 	*x = Headers_Header{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_viz_proto_msgTypes[37]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_viz_proto_msgTypes[37]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *Headers_Header) String() string {
@@ -2798,7 +2722,7 @@ func (*Headers_Header) ProtoMessage() {}
 
 func (x *Headers_Header) ProtoReflect() protoreflect.Message {
 	mi := &file_viz_proto_msgTypes[37]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -2870,11 +2794,9 @@ type PodErrors_PodError struct {
 
 func (x *PodErrors_PodError) Reset() {
 	*x = PodErrors_PodError{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_viz_proto_msgTypes[38]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_viz_proto_msgTypes[38]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *PodErrors_PodError) String() string {
@@ -2885,7 +2807,7 @@ func (*PodErrors_PodError) ProtoMessage() {}
 
 func (x *PodErrors_PodError) ProtoReflect() protoreflect.Message {
 	mi := &file_viz_proto_msgTypes[38]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -2938,11 +2860,9 @@ type PodErrors_PodError_ContainerError struct {
 
 func (x *PodErrors_PodError_ContainerError) Reset() {
 	*x = PodErrors_PodError_ContainerError{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_viz_proto_msgTypes[39]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_viz_proto_msgTypes[39]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *PodErrors_PodError_ContainerError) String() string {
@@ -2953,7 +2873,7 @@ func (*PodErrors_PodError_ContainerError) ProtoMessage() {}
 
 func (x *PodErrors_PodError_ContainerError) ProtoReflect() protoreflect.Message {
 	mi := &file_viz_proto_msgTypes[39]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -3006,11 +2926,9 @@ type StatSummaryResponse_Ok struct {
 
 func (x *StatSummaryResponse_Ok) Reset() {
 	*x = StatSummaryResponse_Ok{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_viz_proto_msgTypes[40]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_viz_proto_msgTypes[40]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *StatSummaryResponse_Ok) String() string {
@@ -3021,7 +2939,7 @@ func (*StatSummaryResponse_Ok) ProtoMessage() {}
 
 func (x *StatSummaryResponse_Ok) ProtoReflect() protoreflect.Message {
 	mi := &file_viz_proto_msgTypes[40]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -3053,11 +2971,9 @@ type AuthzResponse_Ok struct {
 
 func (x *AuthzResponse_Ok) Reset() {
 	*x = AuthzResponse_Ok{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_viz_proto_msgTypes[41]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_viz_proto_msgTypes[41]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *AuthzResponse_Ok) String() string {
@@ -3068,7 +2984,7 @@ func (*AuthzResponse_Ok) ProtoMessage() {}
 
 func (x *AuthzResponse_Ok) ProtoReflect() protoreflect.Message {
 	mi := &file_viz_proto_msgTypes[41]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -3100,11 +3016,9 @@ type StatTable_PodGroup struct {
 
 func (x *StatTable_PodGroup) Reset() {
 	*x = StatTable_PodGroup{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_viz_proto_msgTypes[42]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_viz_proto_msgTypes[42]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *StatTable_PodGroup) String() string {
@@ -3115,7 +3029,7 @@ func (*StatTable_PodGroup) ProtoMessage() {}
 
 func (x *StatTable_PodGroup) ProtoReflect() protoreflect.Message {
 	mi := &file_viz_proto_msgTypes[42]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -3162,11 +3076,9 @@ type StatTable_PodGroup_Row struct {
 
 func (x *StatTable_PodGroup_Row) Reset() {
 	*x = StatTable_PodGroup_Row{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_viz_proto_msgTypes[43]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_viz_proto_msgTypes[43]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *StatTable_PodGroup_Row) String() string {
@@ -3177,7 +3089,7 @@ func (*StatTable_PodGroup_Row) ProtoMessage() {}
 
 func (x *StatTable_PodGroup_Row) ProtoReflect() protoreflect.Message {
 	mi := &file_viz_proto_msgTypes[43]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -3279,11 +3191,9 @@ type EdgesResponse_Ok struct {
 
 func (x *EdgesResponse_Ok) Reset() {
 	*x = EdgesResponse_Ok{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_viz_proto_msgTypes[45]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_viz_proto_msgTypes[45]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *EdgesResponse_Ok) String() string {
@@ -3294,7 +3204,7 @@ func (*EdgesResponse_Ok) ProtoMessage() {}
 
 func (x *EdgesResponse_Ok) ProtoReflect() protoreflect.Message {
 	mi := &file_viz_proto_msgTypes[45]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -3326,11 +3236,9 @@ type TopRoutesResponse_Ok struct {
 
 func (x *TopRoutesResponse_Ok) Reset() {
 	*x = TopRoutesResponse_Ok{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_viz_proto_msgTypes[46]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_viz_proto_msgTypes[46]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *TopRoutesResponse_Ok) String() string {
@@ -3341,7 +3249,7 @@ func (*TopRoutesResponse_Ok) ProtoMessage() {}
 
 func (x *TopRoutesResponse_Ok) ProtoReflect() protoreflect.Message {
 	mi := &file_viz_proto_msgTypes[46]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -3376,11 +3284,9 @@ type RouteTable_Row struct {
 
 func (x *RouteTable_Row) Reset() {
 	*x = RouteTable_Row{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_viz_proto_msgTypes[47]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_viz_proto_msgTypes[47]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *RouteTable_Row) String() string {
@@ -3391,7 +3297,7 @@ func (*RouteTable_Row) ProtoMessage() {}
 
 func (x *RouteTable_Row) ProtoReflect() protoreflect.Message {
 	mi := &file_viz_proto_msgTypes[47]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -3451,11 +3357,9 @@ type GatewaysTable_Row struct {
 
 func (x *GatewaysTable_Row) Reset() {
 	*x = GatewaysTable_Row{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_viz_proto_msgTypes[48]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_viz_proto_msgTypes[48]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *GatewaysTable_Row) String() string {
@@ -3466,7 +3370,7 @@ func (*GatewaysTable_Row) ProtoMessage() {}
 
 func (x *GatewaysTable_Row) ProtoReflect() protoreflect.Message {
 	mi := &file_viz_proto_msgTypes[48]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -3547,11 +3451,9 @@ type GatewaysResponse_Ok struct {
 
 func (x *GatewaysResponse_Ok) Reset() {
 	*x = GatewaysResponse_Ok{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_viz_proto_msgTypes[49]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_viz_proto_msgTypes[49]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *GatewaysResponse_Ok) String() string {
@@ -3562,7 +3464,7 @@ func (*GatewaysResponse_Ok) ProtoMessage() {}
 
 func (x *GatewaysResponse_Ok) ProtoReflect() protoreflect.Message {
 	mi := &file_viz_proto_msgTypes[49]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -4077,7 +3979,7 @@ func file_viz_proto_rawDescGZIP() []byte {
 
 var file_viz_proto_enumTypes = make([]protoimpl.EnumInfo, 3)
 var file_viz_proto_msgTypes = make([]protoimpl.MessageInfo, 50)
-var file_viz_proto_goTypes = []interface{}{
+var file_viz_proto_goTypes = []any{
 	(CheckStatus)(0),                          // 0: linkerd2.viz.CheckStatus
 	(HttpMethod_Registered)(0),                // 1: linkerd2.viz.HttpMethod.Registered
 	(Scheme_Registered)(0),                    // 2: linkerd2.viz.Scheme.Registered
@@ -4217,597 +4119,7 @@ func file_viz_proto_init() {
 	if File_viz_proto != nil {
 		return
 	}
-	if !protoimpl.UnsafeEnabled {
-		file_viz_proto_msgTypes[0].Exporter = func(v interface{}, i int) interface{} {
-			switch v := v.(*Empty); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-		file_viz_proto_msgTypes[1].Exporter = func(v interface{}, i int) interface{} {
-			switch v := v.(*CheckResult); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-		file_viz_proto_msgTypes[2].Exporter = func(v interface{}, i int) interface{} {
-			switch v := v.(*SelfCheckRequest); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-		file_viz_proto_msgTypes[3].Exporter = func(v interface{}, i int) interface{} {
-			switch v := v.(*SelfCheckResponse); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-		file_viz_proto_msgTypes[4].Exporter = func(v interface{}, i int) interface{} {
-			switch v := v.(*ListServicesRequest); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-		file_viz_proto_msgTypes[5].Exporter = func(v interface{}, i int) interface{} {
-			switch v := v.(*ListServicesResponse); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-		file_viz_proto_msgTypes[6].Exporter = func(v interface{}, i int) interface{} {
-			switch v := v.(*Service); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-		file_viz_proto_msgTypes[7].Exporter = func(v interface{}, i int) interface{} {
-			switch v := v.(*ListPodsRequest); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-		file_viz_proto_msgTypes[8].Exporter = func(v interface{}, i int) interface{} {
-			switch v := v.(*ListPodsResponse); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-		file_viz_proto_msgTypes[9].Exporter = func(v interface{}, i int) interface{} {
-			switch v := v.(*Pod); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-		file_viz_proto_msgTypes[10].Exporter = func(v interface{}, i int) interface{} {
-			switch v := v.(*HttpMethod); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-		file_viz_proto_msgTypes[11].Exporter = func(v interface{}, i int) interface{} {
-			switch v := v.(*Scheme); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-		file_viz_proto_msgTypes[12].Exporter = func(v interface{}, i int) interface{} {
-			switch v := v.(*Headers); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-		file_viz_proto_msgTypes[13].Exporter = func(v interface{}, i int) interface{} {
-			switch v := v.(*Eos); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-		file_viz_proto_msgTypes[14].Exporter = func(v interface{}, i int) interface{} {
-			switch v := v.(*ApiError); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-		file_viz_proto_msgTypes[15].Exporter = func(v interface{}, i int) interface{} {
-			switch v := v.(*PodErrors); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-		file_viz_proto_msgTypes[16].Exporter = func(v interface{}, i int) interface{} {
-			switch v := v.(*Resource); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-		file_viz_proto_msgTypes[17].Exporter = func(v interface{}, i int) interface{} {
-			switch v := v.(*ResourceSelection); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-		file_viz_proto_msgTypes[18].Exporter = func(v interface{}, i int) interface{} {
-			switch v := v.(*ResourceError); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-		file_viz_proto_msgTypes[19].Exporter = func(v interface{}, i int) interface{} {
-			switch v := v.(*StatSummaryRequest); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-		file_viz_proto_msgTypes[20].Exporter = func(v interface{}, i int) interface{} {
-			switch v := v.(*StatSummaryResponse); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-		file_viz_proto_msgTypes[21].Exporter = func(v interface{}, i int) interface{} {
-			switch v := v.(*AuthzRequest); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-		file_viz_proto_msgTypes[22].Exporter = func(v interface{}, i int) interface{} {
-			switch v := v.(*AuthzResponse); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-		file_viz_proto_msgTypes[23].Exporter = func(v interface{}, i int) interface{} {
-			switch v := v.(*BasicStats); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-		file_viz_proto_msgTypes[24].Exporter = func(v interface{}, i int) interface{} {
-			switch v := v.(*TcpStats); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-		file_viz_proto_msgTypes[25].Exporter = func(v interface{}, i int) interface{} {
-			switch v := v.(*TrafficSplitStats); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-		file_viz_proto_msgTypes[26].Exporter = func(v interface{}, i int) interface{} {
-			switch v := v.(*ServerStats); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-		file_viz_proto_msgTypes[27].Exporter = func(v interface{}, i int) interface{} {
-			switch v := v.(*StatTable); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-		file_viz_proto_msgTypes[28].Exporter = func(v interface{}, i int) interface{} {
-			switch v := v.(*EdgesRequest); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-		file_viz_proto_msgTypes[29].Exporter = func(v interface{}, i int) interface{} {
-			switch v := v.(*EdgesResponse); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-		file_viz_proto_msgTypes[30].Exporter = func(v interface{}, i int) interface{} {
-			switch v := v.(*Edge); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-		file_viz_proto_msgTypes[31].Exporter = func(v interface{}, i int) interface{} {
-			switch v := v.(*TopRoutesRequest); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-		file_viz_proto_msgTypes[32].Exporter = func(v interface{}, i int) interface{} {
-			switch v := v.(*TopRoutesResponse); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-		file_viz_proto_msgTypes[33].Exporter = func(v interface{}, i int) interface{} {
-			switch v := v.(*RouteTable); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-		file_viz_proto_msgTypes[34].Exporter = func(v interface{}, i int) interface{} {
-			switch v := v.(*GatewaysTable); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-		file_viz_proto_msgTypes[35].Exporter = func(v interface{}, i int) interface{} {
-			switch v := v.(*GatewaysRequest); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-		file_viz_proto_msgTypes[36].Exporter = func(v interface{}, i int) interface{} {
-			switch v := v.(*GatewaysResponse); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-		file_viz_proto_msgTypes[37].Exporter = func(v interface{}, i int) interface{} {
-			switch v := v.(*Headers_Header); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-		file_viz_proto_msgTypes[38].Exporter = func(v interface{}, i int) interface{} {
-			switch v := v.(*PodErrors_PodError); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-		file_viz_proto_msgTypes[39].Exporter = func(v interface{}, i int) interface{} {
-			switch v := v.(*PodErrors_PodError_ContainerError); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-		file_viz_proto_msgTypes[40].Exporter = func(v interface{}, i int) interface{} {
-			switch v := v.(*StatSummaryResponse_Ok); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-		file_viz_proto_msgTypes[41].Exporter = func(v interface{}, i int) interface{} {
-			switch v := v.(*AuthzResponse_Ok); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-		file_viz_proto_msgTypes[42].Exporter = func(v interface{}, i int) interface{} {
-			switch v := v.(*StatTable_PodGroup); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-		file_viz_proto_msgTypes[43].Exporter = func(v interface{}, i int) interface{} {
-			switch v := v.(*StatTable_PodGroup_Row); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-		file_viz_proto_msgTypes[45].Exporter = func(v interface{}, i int) interface{} {
-			switch v := v.(*EdgesResponse_Ok); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-		file_viz_proto_msgTypes[46].Exporter = func(v interface{}, i int) interface{} {
-			switch v := v.(*TopRoutesResponse_Ok); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-		file_viz_proto_msgTypes[47].Exporter = func(v interface{}, i int) interface{} {
-			switch v := v.(*RouteTable_Row); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-		file_viz_proto_msgTypes[48].Exporter = func(v interface{}, i int) interface{} {
-			switch v := v.(*GatewaysTable_Row); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-		file_viz_proto_msgTypes[49].Exporter = func(v interface{}, i int) interface{} {
-			switch v := v.(*GatewaysResponse_Ok); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-	}
-	file_viz_proto_msgTypes[9].OneofWrappers = []interface{}{
+	file_viz_proto_msgTypes[9].OneofWrappers = []any{
 		(*Pod_Deployment)(nil),
 		(*Pod_ReplicaSet)(nil),
 		(*Pod_ReplicationController)(nil),
@@ -4815,55 +4127,55 @@ func file_viz_proto_init() {
 		(*Pod_DaemonSet)(nil),
 		(*Pod_Job)(nil),
 	}
-	file_viz_proto_msgTypes[10].OneofWrappers = []interface{}{
+	file_viz_proto_msgTypes[10].OneofWrappers = []any{
 		(*HttpMethod_Registered_)(nil),
 		(*HttpMethod_Unregistered)(nil),
 	}
-	file_viz_proto_msgTypes[11].OneofWrappers = []interface{}{
+	file_viz_proto_msgTypes[11].OneofWrappers = []any{
 		(*Scheme_Registered_)(nil),
 		(*Scheme_Unregistered)(nil),
 	}
-	file_viz_proto_msgTypes[13].OneofWrappers = []interface{}{
+	file_viz_proto_msgTypes[13].OneofWrappers = []any{
 		(*Eos_GrpcStatusCode)(nil),
 		(*Eos_ResetErrorCode)(nil),
 	}
-	file_viz_proto_msgTypes[19].OneofWrappers = []interface{}{
+	file_viz_proto_msgTypes[19].OneofWrappers = []any{
 		(*StatSummaryRequest_None)(nil),
 		(*StatSummaryRequest_ToResource)(nil),
 		(*StatSummaryRequest_FromResource)(nil),
 	}
-	file_viz_proto_msgTypes[20].OneofWrappers = []interface{}{
+	file_viz_proto_msgTypes[20].OneofWrappers = []any{
 		(*StatSummaryResponse_Ok_)(nil),
 		(*StatSummaryResponse_Error)(nil),
 	}
-	file_viz_proto_msgTypes[22].OneofWrappers = []interface{}{
+	file_viz_proto_msgTypes[22].OneofWrappers = []any{
 		(*AuthzResponse_Ok_)(nil),
 		(*AuthzResponse_Error)(nil),
 	}
-	file_viz_proto_msgTypes[27].OneofWrappers = []interface{}{
+	file_viz_proto_msgTypes[27].OneofWrappers = []any{
 		(*StatTable_PodGroup_)(nil),
 	}
-	file_viz_proto_msgTypes[29].OneofWrappers = []interface{}{
+	file_viz_proto_msgTypes[29].OneofWrappers = []any{
 		(*EdgesResponse_Ok_)(nil),
 		(*EdgesResponse_Error)(nil),
 	}
-	file_viz_proto_msgTypes[31].OneofWrappers = []interface{}{
+	file_viz_proto_msgTypes[31].OneofWrappers = []any{
 		(*TopRoutesRequest_None)(nil),
 		(*TopRoutesRequest_ToResource)(nil),
 	}
-	file_viz_proto_msgTypes[32].OneofWrappers = []interface{}{
+	file_viz_proto_msgTypes[32].OneofWrappers = []any{
 		(*TopRoutesResponse_Error)(nil),
 		(*TopRoutesResponse_Ok_)(nil),
 	}
-	file_viz_proto_msgTypes[36].OneofWrappers = []interface{}{
+	file_viz_proto_msgTypes[36].OneofWrappers = []any{
 		(*GatewaysResponse_Ok_)(nil),
 		(*GatewaysResponse_Error)(nil),
 	}
-	file_viz_proto_msgTypes[37].OneofWrappers = []interface{}{
+	file_viz_proto_msgTypes[37].OneofWrappers = []any{
 		(*Headers_Header_ValueStr)(nil),
 		(*Headers_Header_ValueBin)(nil),
 	}
-	file_viz_proto_msgTypes[38].OneofWrappers = []interface{}{
+	file_viz_proto_msgTypes[38].OneofWrappers = []any{
 		(*PodErrors_PodError_Container)(nil),
 	}
 	type x struct{}
