@@ -1,6 +1,6 @@
 // Code generated by protoc-gen-go. DO NOT EDIT.
 // versions:
-// 	protoc-gen-go v1.28.1
+// 	protoc-gen-go v1.35.2
 // 	protoc        v3.20.0
 // source: viz_tap.proto
 
@@ -72,7 +72,7 @@ func (TapEvent_ProxyDirection) EnumDescriptor() ([]byte, []int) {
 	return file_viz_tap_proto_rawDescGZIP(), []int{2, 0}
 }
 
-// Deprecated: Do not use.
+// Deprecated: Marked as deprecated in viz_tap.proto.
 type TapRequest struct {
 	state         protoimpl.MessageState
 	sizeCache     protoimpl.SizeCache
@@ -97,11 +97,9 @@ type TapRequest struct {
 
 func (x *TapRequest) Reset() {
 	*x = TapRequest{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_viz_tap_proto_msgTypes[0]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_viz_tap_proto_msgTypes[0]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *TapRequest) String() string {
@@ -112,7 +110,7 @@ func (*TapRequest) ProtoMessage() {}
 
 func (x *TapRequest) ProtoReflect() protoreflect.Message {
 	mi := &file_viz_tap_proto_msgTypes[0]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -248,11 +246,9 @@ type TapByResourceRequest struct {
 
 func (x *TapByResourceRequest) Reset() {
 	*x = TapByResourceRequest{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_viz_tap_proto_msgTypes[1]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_viz_tap_proto_msgTypes[1]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *TapByResourceRequest) String() string {
@@ -263,7 +259,7 @@ func (*TapByResourceRequest) ProtoMessage() {}
 
 func (x *TapByResourceRequest) ProtoReflect() protoreflect.Message {
 	mi := &file_viz_tap_proto_msgTypes[1]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -326,11 +322,9 @@ type TapEvent struct {
 
 func (x *TapEvent) Reset() {
 	*x = TapEvent{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_viz_tap_proto_msgTypes[2]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_viz_tap_proto_msgTypes[2]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *TapEvent) String() string {
@@ -341,7 +335,7 @@ func (*TapEvent) ProtoMessage() {}
 
 func (x *TapEvent) ProtoReflect() protoreflect.Message {
 	mi := &file_viz_tap_proto_msgTypes[2]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -439,11 +433,9 @@ type TapByResourceRequest_Match struct {
 
 func (x *TapByResourceRequest_Match) Reset() {
 	*x = TapByResourceRequest_Match{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_viz_tap_proto_msgTypes[3]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_viz_tap_proto_msgTypes[3]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *TapByResourceRequest_Match) String() string {
@@ -454,7 +446,7 @@ func (*TapByResourceRequest_Match) ProtoMessage() {}
 
 func (x *TapByResourceRequest_Match) ProtoReflect() protoreflect.Message {
 	mi := &file_viz_tap_proto_msgTypes[3]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -563,11 +555,9 @@ type TapByResourceRequest_Extract struct {
 
 func (x *TapByResourceRequest_Extract) Reset() {
 	*x = TapByResourceRequest_Extract{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_viz_tap_proto_msgTypes[4]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_viz_tap_proto_msgTypes[4]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *TapByResourceRequest_Extract) String() string {
@@ -578,7 +568,7 @@ func (*TapByResourceRequest_Extract) ProtoMessage() {}
 
 func (x *TapByResourceRequest_Extract) ProtoReflect() protoreflect.Message {
 	mi := &file_viz_tap_proto_msgTypes[4]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -627,11 +617,9 @@ type TapByResourceRequest_Match_Seq struct {
 
 func (x *TapByResourceRequest_Match_Seq) Reset() {
 	*x = TapByResourceRequest_Match_Seq{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_viz_tap_proto_msgTypes[5]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_viz_tap_proto_msgTypes[5]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *TapByResourceRequest_Match_Seq) String() string {
@@ -642,7 +630,7 @@ func (*TapByResourceRequest_Match_Seq) ProtoMessage() {}
 
 func (x *TapByResourceRequest_Match_Seq) ProtoReflect() protoreflect.Message {
 	mi := &file_viz_tap_proto_msgTypes[5]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -680,11 +668,9 @@ type TapByResourceRequest_Match_Http struct {
 
 func (x *TapByResourceRequest_Match_Http) Reset() {
 	*x = TapByResourceRequest_Match_Http{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_viz_tap_proto_msgTypes[6]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_viz_tap_proto_msgTypes[6]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *TapByResourceRequest_Match_Http) String() string {
@@ -695,7 +681,7 @@ func (*TapByResourceRequest_Match_Http) ProtoMessage() {}
 
 func (x *TapByResourceRequest_Match_Http) ProtoReflect() protoreflect.Message {
 	mi := &file_viz_tap_proto_msgTypes[6]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -786,11 +772,9 @@ type TapByResourceRequest_Extract_Http struct {
 
 func (x *TapByResourceRequest_Extract_Http) Reset() {
 	*x = TapByResourceRequest_Extract_Http{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_viz_tap_proto_msgTypes[7]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_viz_tap_proto_msgTypes[7]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *TapByResourceRequest_Extract_Http) String() string {
@@ -801,7 +785,7 @@ func (*TapByResourceRequest_Extract_Http) ProtoMessage() {}
 
 func (x *TapByResourceRequest_Extract_Http) ProtoReflect() protoreflect.Message {
 	mi := &file_viz_tap_proto_msgTypes[7]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -848,11 +832,9 @@ type TapByResourceRequest_Extract_Http_Headers struct {
 
 func (x *TapByResourceRequest_Extract_Http_Headers) Reset() {
 	*x = TapByResourceRequest_Extract_Http_Headers{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_viz_tap_proto_msgTypes[8]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_viz_tap_proto_msgTypes[8]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *TapByResourceRequest_Extract_Http_Headers) String() string {
@@ -863,7 +845,7 @@ func (*TapByResourceRequest_Extract_Http_Headers) ProtoMessage() {}
 
 func (x *TapByResourceRequest_Extract_Http_Headers) ProtoReflect() protoreflect.Message {
 	mi := &file_viz_tap_proto_msgTypes[8]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -888,11 +870,9 @@ type TapEvent_EndpointMeta struct {
 
 func (x *TapEvent_EndpointMeta) Reset() {
 	*x = TapEvent_EndpointMeta{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_viz_tap_proto_msgTypes[9]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_viz_tap_proto_msgTypes[9]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *TapEvent_EndpointMeta) String() string {
@@ -903,7 +883,7 @@ func (*TapEvent_EndpointMeta) ProtoMessage() {}
 
 func (x *TapEvent_EndpointMeta) ProtoReflect() protoreflect.Message {
 	mi := &file_viz_tap_proto_msgTypes[9]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -935,11 +915,9 @@ type TapEvent_RouteMeta struct {
 
 func (x *TapEvent_RouteMeta) Reset() {
 	*x = TapEvent_RouteMeta{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_viz_tap_proto_msgTypes[10]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_viz_tap_proto_msgTypes[10]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *TapEvent_RouteMeta) String() string {
@@ -950,7 +928,7 @@ func (*TapEvent_RouteMeta) ProtoMessage() {}
 
 func (x *TapEvent_RouteMeta) ProtoReflect() protoreflect.Message {
 	mi := &file_viz_tap_proto_msgTypes[10]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -987,11 +965,9 @@ type TapEvent_Http struct {
 
 func (x *TapEvent_Http) Reset() {
 	*x = TapEvent_Http{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_viz_tap_proto_msgTypes[11]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_viz_tap_proto_msgTypes[11]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *TapEvent_Http) String() string {
@@ -1002,7 +978,7 @@ func (*TapEvent_Http) ProtoMessage() {}
 
 func (x *TapEvent_Http) ProtoReflect() protoreflect.Message {
 	mi := &file_viz_tap_proto_msgTypes[11]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -1080,11 +1056,9 @@ type TapEvent_Http_StreamId struct {
 
 func (x *TapEvent_Http_StreamId) Reset() {
 	*x = TapEvent_Http_StreamId{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_viz_tap_proto_msgTypes[14]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_viz_tap_proto_msgTypes[14]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *TapEvent_Http_StreamId) String() string {
@@ -1095,7 +1069,7 @@ func (*TapEvent_Http_StreamId) ProtoMessage() {}
 
 func (x *TapEvent_Http_StreamId) ProtoReflect() protoreflect.Message {
 	mi := &file_viz_tap_proto_msgTypes[14]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -1139,11 +1113,9 @@ type TapEvent_Http_RequestInit struct {
 
 func (x *TapEvent_Http_RequestInit) Reset() {
 	*x = TapEvent_Http_RequestInit{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_viz_tap_proto_msgTypes[15]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_viz_tap_proto_msgTypes[15]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *TapEvent_Http_RequestInit) String() string {
@@ -1154,7 +1126,7 @@ func (*TapEvent_Http_RequestInit) ProtoMessage() {}
 
 func (x *TapEvent_Http_RequestInit) ProtoReflect() protoreflect.Message {
 	mi := &file_viz_tap_proto_msgTypes[15]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -1224,11 +1196,9 @@ type TapEvent_Http_ResponseInit struct {
 
 func (x *TapEvent_Http_ResponseInit) Reset() {
 	*x = TapEvent_Http_ResponseInit{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_viz_tap_proto_msgTypes[16]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_viz_tap_proto_msgTypes[16]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *TapEvent_Http_ResponseInit) String() string {
@@ -1239,7 +1209,7 @@ func (*TapEvent_Http_ResponseInit) ProtoMessage() {}
 
 func (x *TapEvent_Http_ResponseInit) ProtoReflect() protoreflect.Message {
 	mi := &file_viz_tap_proto_msgTypes[16]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -1297,11 +1267,9 @@ type TapEvent_Http_ResponseEnd struct {
 
 func (x *TapEvent_Http_ResponseEnd) Reset() {
 	*x = TapEvent_Http_ResponseEnd{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_viz_tap_proto_msgTypes[17]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_viz_tap_proto_msgTypes[17]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *TapEvent_Http_ResponseEnd) String() string {
@@ -1312,7 +1280,7 @@ func (*TapEvent_Http_ResponseEnd) ProtoMessage() {}
 
 func (x *TapEvent_Http_ResponseEnd) ProtoReflect() protoreflect.Message {
 	mi := &file_viz_tap_proto_msgTypes[17]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -1613,7 +1581,7 @@ func file_viz_tap_proto_rawDescGZIP() []byte {
 
 var file_viz_tap_proto_enumTypes = make([]protoimpl.EnumInfo, 1)
 var file_viz_tap_proto_msgTypes = make([]protoimpl.MessageInfo, 18)
-var file_viz_tap_proto_goTypes = []interface{}{
+var file_viz_tap_proto_goTypes = []any{
 	(TapEvent_ProxyDirection)(0),                      // 0: linkerd2.tap.TapEvent.ProxyDirection
 	(*TapRequest)(nil),                                // 1: linkerd2.tap.TapRequest
 	(*TapByResourceRequest)(nil),                      // 2: linkerd2.tap.TapByResourceRequest
@@ -1693,227 +1661,33 @@ func file_viz_tap_proto_init() {
 	if File_viz_tap_proto != nil {
 		return
 	}
-	if !protoimpl.UnsafeEnabled {
-		file_viz_tap_proto_msgTypes[0].Exporter = func(v interface{}, i int) interface{} {
-			switch v := v.(*TapRequest); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-		file_viz_tap_proto_msgTypes[1].Exporter = func(v interface{}, i int) interface{} {
-			switch v := v.(*TapByResourceRequest); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-		file_viz_tap_proto_msgTypes[2].Exporter = func(v interface{}, i int) interface{} {
-			switch v := v.(*TapEvent); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-		file_viz_tap_proto_msgTypes[3].Exporter = func(v interface{}, i int) interface{} {
-			switch v := v.(*TapByResourceRequest_Match); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-		file_viz_tap_proto_msgTypes[4].Exporter = func(v interface{}, i int) interface{} {
-			switch v := v.(*TapByResourceRequest_Extract); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-		file_viz_tap_proto_msgTypes[5].Exporter = func(v interface{}, i int) interface{} {
-			switch v := v.(*TapByResourceRequest_Match_Seq); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-		file_viz_tap_proto_msgTypes[6].Exporter = func(v interface{}, i int) interface{} {
-			switch v := v.(*TapByResourceRequest_Match_Http); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-		file_viz_tap_proto_msgTypes[7].Exporter = func(v interface{}, i int) interface{} {
-			switch v := v.(*TapByResourceRequest_Extract_Http); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-		file_viz_tap_proto_msgTypes[8].Exporter = func(v interface{}, i int) interface{} {
-			switch v := v.(*TapByResourceRequest_Extract_Http_Headers); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-		file_viz_tap_proto_msgTypes[9].Exporter = func(v interface{}, i int) interface{} {
-			switch v := v.(*TapEvent_EndpointMeta); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-		file_viz_tap_proto_msgTypes[10].Exporter = func(v interface{}, i int) interface{} {
-			switch v := v.(*TapEvent_RouteMeta); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-		file_viz_tap_proto_msgTypes[11].Exporter = func(v interface{}, i int) interface{} {
-			switch v := v.(*TapEvent_Http); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-		file_viz_tap_proto_msgTypes[14].Exporter = func(v interface{}, i int) interface{} {
-			switch v := v.(*TapEvent_Http_StreamId); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-		file_viz_tap_proto_msgTypes[15].Exporter = func(v interface{}, i int) interface{} {
-			switch v := v.(*TapEvent_Http_RequestInit); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-		file_viz_tap_proto_msgTypes[16].Exporter = func(v interface{}, i int) interface{} {
-			switch v := v.(*TapEvent_Http_ResponseInit); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-		file_viz_tap_proto_msgTypes[17].Exporter = func(v interface{}, i int) interface{} {
-			switch v := v.(*TapEvent_Http_ResponseEnd); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-	}
-	file_viz_tap_proto_msgTypes[0].OneofWrappers = []interface{}{
+	file_viz_tap_proto_msgTypes[0].OneofWrappers = []any{
 		(*TapRequest_Pod)(nil),
 		(*TapRequest_Deployment)(nil),
 	}
-	file_viz_tap_proto_msgTypes[2].OneofWrappers = []interface{}{
+	file_viz_tap_proto_msgTypes[2].OneofWrappers = []any{
 		(*TapEvent_Http_)(nil),
 	}
-	file_viz_tap_proto_msgTypes[3].OneofWrappers = []interface{}{
+	file_viz_tap_proto_msgTypes[3].OneofWrappers = []any{
 		(*TapByResourceRequest_Match_All)(nil),
 		(*TapByResourceRequest_Match_Any)(nil),
 		(*TapByResourceRequest_Match_Not)(nil),
 		(*TapByResourceRequest_Match_Destinations)(nil),
 		(*TapByResourceRequest_Match_Http_)(nil),
 	}
-	file_viz_tap_proto_msgTypes[4].OneofWrappers = []interface{}{
+	file_viz_tap_proto_msgTypes[4].OneofWrappers = []any{
 		(*TapByResourceRequest_Extract_Http_)(nil),
 	}
-	file_viz_tap_proto_msgTypes[6].OneofWrappers = []interface{}{
+	file_viz_tap_proto_msgTypes[6].OneofWrappers = []any{
 		(*TapByResourceRequest_Match_Http_Scheme)(nil),
 		(*TapByResourceRequest_Match_Http_Method)(nil),
 		(*TapByResourceRequest_Match_Http_Authority)(nil),
 		(*TapByResourceRequest_Match_Http_Path)(nil),
 	}
-	file_viz_tap_proto_msgTypes[7].OneofWrappers = []interface{}{
+	file_viz_tap_proto_msgTypes[7].OneofWrappers = []any{
 		(*TapByResourceRequest_Extract_Http_Headers_)(nil),
 	}
-	file_viz_tap_proto_msgTypes[11].OneofWrappers = []interface{}{
+	file_viz_tap_proto_msgTypes[11].OneofWrappers = []any{
 		(*TapEvent_Http_RequestInit_)(nil),
 		(*TapEvent_Http_ResponseInit_)(nil),
 		(*TapEvent_Http_ResponseEnd_)(nil),
