@@ -72,9 +72,6 @@ fn convert_outbound_route(
     parent_info: &ParentInfo,
     original_dst: Option<SocketAddr>,
 ) -> outbound::GrpcRoute {
-    // This encoder sets deprecated timeouts for older proxies.
-    #![allow(deprecated)]
-
     let metadata = Some(meta::Metadata {
         kind: Some(meta::metadata::Kind::Resource(meta::Resource {
             group: gknn.group.to_string(),
@@ -122,6 +119,7 @@ fn convert_outbound_route(
                 if retry.is_none() {
                     retry = service_retry.clone();
                 }
+                #[allow(deprecated)]
                 outbound::grpc_route::Rule {
                     matches: matches.into_iter().map(convert_match).collect(),
                     backends: Some(outbound::grpc_route::Distribution { kind: Some(dist) }),
