@@ -1,6 +1,6 @@
 // Code generated by protoc-gen-go-grpc. DO NOT EDIT.
 // versions:
-// - protoc-gen-go-grpc v1.2.0
+// - protoc-gen-go-grpc v1.5.1
 // - protoc             v3.20.0
 // source: viz.proto
 
@@ -15,8 +15,19 @@ import (
 
 // This is a compile-time assertion to ensure that this generated file
 // is compatible with the grpc package it is being compiled against.
-// Requires gRPC-Go v1.32.0 or later.
-const _ = grpc.SupportPackageIsVersion7
+// Requires gRPC-Go v1.64.0 or later.
+const _ = grpc.SupportPackageIsVersion9
+
+const (
+	Api_StatSummary_FullMethodName  = "/linkerd2.viz.Api/StatSummary"
+	Api_Edges_FullMethodName        = "/linkerd2.viz.Api/Edges"
+	Api_Gateways_FullMethodName     = "/linkerd2.viz.Api/Gateways"
+	Api_TopRoutes_FullMethodName    = "/linkerd2.viz.Api/TopRoutes"
+	Api_ListPods_FullMethodName     = "/linkerd2.viz.Api/ListPods"
+	Api_ListServices_FullMethodName = "/linkerd2.viz.Api/ListServices"
+	Api_SelfCheck_FullMethodName    = "/linkerd2.viz.Api/SelfCheck"
+	Api_Authz_FullMethodName        = "/linkerd2.viz.Api/Authz"
+)
 
 // ApiClient is the client API for Api service.
 //
@@ -41,8 +52,9 @@ func NewApiClient(cc grpc.ClientConnInterface) ApiClient {
 }
 
 func (c *apiClient) StatSummary(ctx context.Context, in *StatSummaryRequest, opts ...grpc.CallOption) (*StatSummaryResponse, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
 	out := new(StatSummaryResponse)
-	err := c.cc.Invoke(ctx, "/linkerd2.viz.Api/StatSummary", in, out, opts...)
+	err := c.cc.Invoke(ctx, Api_StatSummary_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
@@ -50,8 +62,9 @@ func (c *apiClient) StatSummary(ctx context.Context, in *StatSummaryRequest, opt
 }
 
 func (c *apiClient) Edges(ctx context.Context, in *EdgesRequest, opts ...grpc.CallOption) (*EdgesResponse, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
 	out := new(EdgesResponse)
-	err := c.cc.Invoke(ctx, "/linkerd2.viz.Api/Edges", in, out, opts...)
+	err := c.cc.Invoke(ctx, Api_Edges_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
@@ -59,8 +72,9 @@ func (c *apiClient) Edges(ctx context.Context, in *EdgesRequest, opts ...grpc.Ca
 }
 
 func (c *apiClient) Gateways(ctx context.Context, in *GatewaysRequest, opts ...grpc.CallOption) (*GatewaysResponse, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
 	out := new(GatewaysResponse)
-	err := c.cc.Invoke(ctx, "/linkerd2.viz.Api/Gateways", in, out, opts...)
+	err := c.cc.Invoke(ctx, Api_Gateways_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
@@ -68,8 +82,9 @@ func (c *apiClient) Gateways(ctx context.Context, in *GatewaysRequest, opts ...g
 }
 
 func (c *apiClient) TopRoutes(ctx context.Context, in *TopRoutesRequest, opts ...grpc.CallOption) (*TopRoutesResponse, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
 	out := new(TopRoutesResponse)
-	err := c.cc.Invoke(ctx, "/linkerd2.viz.Api/TopRoutes", in, out, opts...)
+	err := c.cc.Invoke(ctx, Api_TopRoutes_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
@@ -77,8 +92,9 @@ func (c *apiClient) TopRoutes(ctx context.Context, in *TopRoutesRequest, opts ..
 }
 
 func (c *apiClient) ListPods(ctx context.Context, in *ListPodsRequest, opts ...grpc.CallOption) (*ListPodsResponse, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
 	out := new(ListPodsResponse)
-	err := c.cc.Invoke(ctx, "/linkerd2.viz.Api/ListPods", in, out, opts...)
+	err := c.cc.Invoke(ctx, Api_ListPods_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
@@ -86,8 +102,9 @@ func (c *apiClient) ListPods(ctx context.Context, in *ListPodsRequest, opts ...g
 }
 
 func (c *apiClient) ListServices(ctx context.Context, in *ListServicesRequest, opts ...grpc.CallOption) (*ListServicesResponse, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
 	out := new(ListServicesResponse)
-	err := c.cc.Invoke(ctx, "/linkerd2.viz.Api/ListServices", in, out, opts...)
+	err := c.cc.Invoke(ctx, Api_ListServices_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
@@ -95,8 +112,9 @@ func (c *apiClient) ListServices(ctx context.Context, in *ListServicesRequest, o
 }
 
 func (c *apiClient) SelfCheck(ctx context.Context, in *SelfCheckRequest, opts ...grpc.CallOption) (*SelfCheckResponse, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
 	out := new(SelfCheckResponse)
-	err := c.cc.Invoke(ctx, "/linkerd2.viz.Api/SelfCheck", in, out, opts...)
+	err := c.cc.Invoke(ctx, Api_SelfCheck_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
@@ -104,8 +122,9 @@ func (c *apiClient) SelfCheck(ctx context.Context, in *SelfCheckRequest, opts ..
 }
 
 func (c *apiClient) Authz(ctx context.Context, in *AuthzRequest, opts ...grpc.CallOption) (*AuthzResponse, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
 	out := new(AuthzResponse)
-	err := c.cc.Invoke(ctx, "/linkerd2.viz.Api/Authz", in, out, opts...)
+	err := c.cc.Invoke(ctx, Api_Authz_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
@@ -114,7 +133,7 @@ func (c *apiClient) Authz(ctx context.Context, in *AuthzRequest, opts ...grpc.Ca
 
 // ApiServer is the server API for Api service.
 // All implementations must embed UnimplementedApiServer
-// for forward compatibility
+// for forward compatibility.
 type ApiServer interface {
 	StatSummary(context.Context, *StatSummaryRequest) (*StatSummaryResponse, error)
 	Edges(context.Context, *EdgesRequest) (*EdgesResponse, error)
@@ -127,9 +146,12 @@ type ApiServer interface {
 	mustEmbedUnimplementedApiServer()
 }
 
-// UnimplementedApiServer must be embedded to have forward compatible implementations.
-type UnimplementedApiServer struct {
-}
+// UnimplementedApiServer must be embedded to have
+// forward compatible implementations.
+//
+// NOTE: this should be embedded by value instead of pointer to avoid a nil
+// pointer dereference when methods are called.
+type UnimplementedApiServer struct{}
 
 func (UnimplementedApiServer) StatSummary(context.Context, *StatSummaryRequest) (*StatSummaryResponse, error) {
 	return nil, status.Errorf(codes.Unimplemented, "method StatSummary not implemented")
@@ -156,6 +178,7 @@ func (UnimplementedApiServer) Authz(context.Context, *AuthzRequest) (*AuthzRespo
 	return nil, status.Errorf(codes.Unimplemented, "method Authz not implemented")
 }
 func (UnimplementedApiServer) mustEmbedUnimplementedApiServer() {}
+func (UnimplementedApiServer) testEmbeddedByValue()             {}
 
 // UnsafeApiServer may be embedded to opt out of forward compatibility for this service.
 // Use of this interface is not recommended, as added methods to ApiServer will
@@ -165,6 +188,13 @@ type UnsafeApiServer interface {
 }
 
 func RegisterApiServer(s grpc.ServiceRegistrar, srv ApiServer) {
+	// If the following call pancis, it indicates UnimplementedApiServer was
+	// embedded by pointer and is nil.  This will cause panics if an
+	// unimplemented method is ever invoked, so we test this at initialization
+	// time to prevent it from happening at runtime later due to I/O.
+	if t, ok := srv.(interface{ testEmbeddedByValue() }); ok {
+		t.testEmbeddedByValue()
+	}
 	s.RegisterService(&Api_ServiceDesc, srv)
 }
 
@@ -178,7 +208,7 @@ func _Api_StatSummary_Handler(srv interface{}, ctx context.Context, dec func(int
 	}
 	info := &grpc.UnaryServerInfo{
 		Server:     srv,
-		FullMethod: "/linkerd2.viz.Api/StatSummary",
+		FullMethod: Api_StatSummary_FullMethodName,
 	}
 	handler := func(ctx context.Context, req interface{}) (interface{}, error) {
 		return srv.(ApiServer).StatSummary(ctx, req.(*StatSummaryRequest))
@@ -196,7 +226,7 @@ func _Api_Edges_Handler(srv interface{}, ctx context.Context, dec func(interface
 	}
 	info := &grpc.UnaryServerInfo{
 		Server:     srv,
-		FullMethod: "/linkerd2.viz.Api/Edges",
+		FullMethod: Api_Edges_FullMethodName,
 	}
 	handler := func(ctx context.Context, req interface{}) (interface{}, error) {
 		return srv.(ApiServer).Edges(ctx, req.(*EdgesRequest))
@@ -214,7 +244,7 @@ func _Api_Gateways_Handler(srv interface{}, ctx context.Context, dec func(interf
 	}
 	info := &grpc.UnaryServerInfo{
 		Server:     srv,
-		FullMethod: "/linkerd2.viz.Api/Gateways",
+		FullMethod: Api_Gateways_FullMethodName,
 	}
 	handler := func(ctx context.Context, req interface{}) (interface{}, error) {
 		return srv.(ApiServer).Gateways(ctx, req.(*GatewaysRequest))
@@ -232,7 +262,7 @@ func _Api_TopRoutes_Handler(srv interface{}, ctx context.Context, dec func(inter
 	}
 	info := &grpc.UnaryServerInfo{
 		Server:     srv,
-		FullMethod: "/linkerd2.viz.Api/TopRoutes",
+		FullMethod: Api_TopRoutes_FullMethodName,
 	}
 	handler := func(ctx context.Context, req interface{}) (interface{}, error) {
 		return srv.(ApiServer).TopRoutes(ctx, req.(*TopRoutesRequest))
@@ -250,7 +280,7 @@ func _Api_ListPods_Handler(srv interface{}, ctx context.Context, dec func(interf
 	}
 	info := &grpc.UnaryServerInfo{
 		Server:     srv,
-		FullMethod: "/linkerd2.viz.Api/ListPods",
+		FullMethod: Api_ListPods_FullMethodName,
 	}
 	handler := func(ctx context.Context, req interface{}) (interface{}, error) {
 		return srv.(ApiServer).ListPods(ctx, req.(*ListPodsRequest))
@@ -268,7 +298,7 @@ func _Api_ListServices_Handler(srv interface{}, ctx context.Context, dec func(in
 	}
 	info := &grpc.UnaryServerInfo{
 		Server:     srv,
-		FullMethod: "/linkerd2.viz.Api/ListServices",
+		FullMethod: Api_ListServices_FullMethodName,
 	}
 	handler := func(ctx context.Context, req interface{}) (interface{}, error) {
 		return srv.(ApiServer).ListServices(ctx, req.(*ListServicesRequest))
@@ -286,7 +316,7 @@ func _Api_SelfCheck_Handler(srv interface{}, ctx context.Context, dec func(inter
 	}
 	info := &grpc.UnaryServerInfo{
 		Server:     srv,
-		FullMethod: "/linkerd2.viz.Api/SelfCheck",
+		FullMethod: Api_SelfCheck_FullMethodName,
 	}
 	handler := func(ctx context.Context, req interface{}) (interface{}, error) {
 		return srv.(ApiServer).SelfCheck(ctx, req.(*SelfCheckRequest))
@@ -304,7 +334,7 @@ func _Api_Authz_Handler(srv interface{}, ctx context.Context, dec func(interface
 	}
 	info := &grpc.UnaryServerInfo{
 		Server:     srv,
-		FullMethod: "/linkerd2.viz.Api/Authz",
+		FullMethod: Api_Authz_FullMethodName,
 	}
 	handler := func(ctx context.Context, req interface{}) (interface{}, error) {
 		return srv.(ApiServer).Authz(ctx, req.(*AuthzRequest))
