@@ -1,6 +1,6 @@
 // Code generated by protoc-gen-go-grpc. DO NOT EDIT.
 // versions:
-// - protoc-gen-go-grpc v1.2.0
+// - protoc-gen-go-grpc v1.5.1
 // - protoc             v3.20.0
 // source: viz_tap.proto
 
@@ -15,17 +15,22 @@ import (
 
 // This is a compile-time assertion to ensure that this generated file
 // is compatible with the grpc package it is being compiled against.
-// Requires gRPC-Go v1.32.0 or later.
-const _ = grpc.SupportPackageIsVersion7
+// Requires gRPC-Go v1.64.0 or later.
+const _ = grpc.SupportPackageIsVersion9
+
+const (
+	Tap_Tap_FullMethodName           = "/linkerd2.tap.Tap/Tap"
+	Tap_TapByResource_FullMethodName = "/linkerd2.tap.Tap/TapByResource"
+)
 
 // TapClient is the client API for Tap service.
 //
 // For semantics around ctx use and closing/ending streaming RPCs, please refer to https://pkg.go.dev/google.golang.org/grpc/?tab=doc#ClientConn.NewStream.
 type TapClient interface {
 	// Deprecated: Do not use.
-	Tap(ctx context.Context, in *TapRequest, opts ...grpc.CallOption) (Tap_TapClient, error)
+	Tap(ctx context.Context, in *TapRequest, opts ...grpc.CallOption) (grpc.ServerStreamingClient[TapEvent], error)
 	// Deprecated: Do not use.
-	TapByResource(ctx context.Context, in *TapByResourceRequest, opts ...grpc.CallOption) (Tap_TapByResourceClient, error)
+	TapByResource(ctx context.Context, in *TapByResourceRequest, opts ...grpc.CallOption) (grpc.ServerStreamingClient[TapEvent], error)
 }
 
 type tapClient struct {
@@ -37,12 +42,13 @@ func NewTapClient(cc grpc.ClientConnInterface) TapClient {
 }
 
 // Deprecated: Do not use.
-func (c *tapClient) Tap(ctx context.Context, in *TapRequest, opts ...grpc.CallOption) (Tap_TapClient, error) {
-	stream, err := c.cc.NewStream(ctx, &Tap_ServiceDesc.Streams[0], "/linkerd2.tap.Tap/Tap", opts...)
+func (c *tapClient) Tap(ctx context.Context, in *TapRequest, opts ...grpc.CallOption) (grpc.ServerStreamingClient[TapEvent], error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
+	stream, err := c.cc.NewStream(ctx, &Tap_ServiceDesc.Streams[0], Tap_Tap_FullMethodName, cOpts...)
 	if err != nil {
 		return nil, err
 	}
-	x := &tapTapClient{stream}
+	x := &grpc.GenericClientStream[TapRequest, TapEvent]{ClientStream: stream}
 	if err := x.ClientStream.SendMsg(in); err != nil {
 		return nil, err
 	}
@@ -52,30 +58,17 @@ func (c *tapClient) Tap(ctx context.Context, in *TapRequest, opts ...grpc.CallOp
 	return x, nil
 }
 
-type Tap_TapClient interface {
-	Recv() (*TapEvent, error)
-	grpc.ClientStream
-}
-
-type tapTapClient struct {
-	grpc.ClientStream
-}
-
-func (x *tapTapClient) Recv() (*TapEvent, error) {
-	m := new(TapEvent)
-	if err := x.ClientStream.RecvMsg(m); err != nil {
-		return nil, err
-	}
-	return m, nil
-}
+// This type alias is provided for backwards compatibility with existing code that references the prior non-generic stream type by name.
+type Tap_TapClient = grpc.ServerStreamingClient[TapEvent]
 
 // Deprecated: Do not use.
-func (c *tapClient) TapByResource(ctx context.Context, in *TapByResourceRequest, opts ...grpc.CallOption) (Tap_TapByResourceClient, error) {
-	stream, err := c.cc.NewStream(ctx, &Tap_ServiceDesc.Streams[1], "/linkerd2.tap.Tap/TapByResource", opts...)
+func (c *tapClient) TapByResource(ctx context.Context, in *TapByResourceRequest, opts ...grpc.CallOption) (grpc.ServerStreamingClient[TapEvent], error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
+	stream, err := c.cc.NewStream(ctx, &Tap_ServiceDesc.Streams[1], Tap_TapByResource_FullMethodName, cOpts...)
 	if err != nil {
 		return nil, err
 	}
-	x := &tapTapByResourceClient{stream}
+	x := &grpc.GenericClientStream[TapByResourceRequest, TapEvent]{ClientStream: stream}
 	if err := x.ClientStream.SendMsg(in); err != nil {
 		return nil, err
 	}
@@ -85,45 +78,35 @@ func (c *tapClient) TapByResource(ctx context.Context, in *TapByResourceRequest,
 	return x, nil
 }
 
-type Tap_TapByResourceClient interface {
-	Recv() (*TapEvent, error)
-	grpc.ClientStream
-}
-
-type tapTapByResourceClient struct {
-	grpc.ClientStream
-}
-
-func (x *tapTapByResourceClient) Recv() (*TapEvent, error) {
-	m := new(TapEvent)
-	if err := x.ClientStream.RecvMsg(m); err != nil {
-		return nil, err
-	}
-	return m, nil
-}
+// This type alias is provided for backwards compatibility with existing code that references the prior non-generic stream type by name.
+type Tap_TapByResourceClient = grpc.ServerStreamingClient[TapEvent]
 
 // TapServer is the server API for Tap service.
 // All implementations must embed UnimplementedTapServer
-// for forward compatibility
+// for forward compatibility.
 type TapServer interface {
 	// Deprecated: Do not use.
-	Tap(*TapRequest, Tap_TapServer) error
+	Tap(*TapRequest, grpc.ServerStreamingServer[TapEvent]) error
 	// Deprecated: Do not use.
-	TapByResource(*TapByResourceRequest, Tap_TapByResourceServer) error
+	TapByResource(*TapByResourceRequest, grpc.ServerStreamingServer[TapEvent]) error
 	mustEmbedUnimplementedTapServer()
 }
 
-// UnimplementedTapServer must be embedded to have forward compatible implementations.
-type UnimplementedTapServer struct {
-}
+// UnimplementedTapServer must be embedded to have
+// forward compatible implementations.
+//
+// NOTE: this should be embedded by value instead of pointer to avoid a nil
+// pointer dereference when methods are called.
+type UnimplementedTapServer struct{}
 
-func (UnimplementedTapServer) Tap(*TapRequest, Tap_TapServer) error {
+func (UnimplementedTapServer) Tap(*TapRequest, grpc.ServerStreamingServer[TapEvent]) error {
 	return status.Errorf(codes.Unimplemented, "method Tap not implemented")
 }
-func (UnimplementedTapServer) TapByResource(*TapByResourceRequest, Tap_TapByResourceServer) error {
+func (UnimplementedTapServer) TapByResource(*TapByResourceRequest, grpc.ServerStreamingServer[TapEvent]) error {
 	return status.Errorf(codes.Unimplemented, "method TapByResource not implemented")
 }
 func (UnimplementedTapServer) mustEmbedUnimplementedTapServer() {}
+func (UnimplementedTapServer) testEmbeddedByValue()             {}
 
 // UnsafeTapServer may be embedded to opt out of forward compatibility for this service.
 // Use of this interface is not recommended, as added methods to TapServer will
@@ -133,6 +116,13 @@ type UnsafeTapServer interface {
 }
 
 func RegisterTapServer(s grpc.ServiceRegistrar, srv TapServer) {
+	// If the following call pancis, it indicates UnimplementedTapServer was
+	// embedded by pointer and is nil.  This will cause panics if an
+	// unimplemented method is ever invoked, so we test this at initialization
+	// time to prevent it from happening at runtime later due to I/O.
+	if t, ok := srv.(interface{ testEmbeddedByValue() }); ok {
+		t.testEmbeddedByValue()
+	}
 	s.RegisterService(&Tap_ServiceDesc, srv)
 }
 
@@ -141,42 +131,22 @@ func _Tap_Tap_Handler(srv interface{}, stream grpc.ServerStream) error {
 	if err := stream.RecvMsg(m); err != nil {
 		return err
 	}
-	return srv.(TapServer).Tap(m, &tapTapServer{stream})
+	return srv.(TapServer).Tap(m, &grpc.GenericServerStream[TapRequest, TapEvent]{ServerStream: stream})
 }
 
-type Tap_TapServer interface {
-	Send(*TapEvent) error
-	grpc.ServerStream
-}
-
-type tapTapServer struct {
-	grpc.ServerStream
-}
-
-func (x *tapTapServer) Send(m *TapEvent) error {
-	return x.ServerStream.SendMsg(m)
-}
+// This type alias is provided for backwards compatibility with existing code that references the prior non-generic stream type by name.
+type Tap_TapServer = grpc.ServerStreamingServer[TapEvent]
 
 func _Tap_TapByResource_Handler(srv interface{}, stream grpc.ServerStream) error {
 	m := new(TapByResourceRequest)
 	if err := stream.RecvMsg(m); err != nil {
 		return err
 	}
-	return srv.(TapServer).TapByResource(m, &tapTapByResourceServer{stream})
+	return srv.(TapServer).TapByResource(m, &grpc.GenericServerStream[TapByResourceRequest, TapEvent]{ServerStream: stream})
 }
 
-type Tap_TapByResourceServer interface {
-	Send(*TapEvent) error
-	grpc.ServerStream
-}
-
-type tapTapByResourceServer struct {
-	grpc.ServerStream
-}
-
-func (x *tapTapByResourceServer) Send(m *TapEvent) error {
-	return x.ServerStream.SendMsg(m)
-}
+// This type alias is provided for backwards compatibility with existing code that references the prior non-generic stream type by name.
+type Tap_TapByResourceServer = grpc.ServerStreamingServer[TapEvent]
 
 // Tap_ServiceDesc is the grpc.ServiceDesc for Tap service.
 // It's only intended for direct use with grpc.RegisterService,
